pyproject.toml
[project]
name = "einstein"
version = "0.1.0"
description = ""
authors = [
    {name = "Kadyr Gullyyev",email = "kadyr.gullyyew@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "aioboto3 (>=14.3.0,<15.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
